web_scraper/requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
numpy==2.1.2
outcome==1.3.0.post0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
